Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -55,7 +55,7 @@ def get_average_age_for_season(data, season)
   end
   total = 0
   ages_array.each do |i|
-    total += i.to_i
+    total += i.to_f
   end
   total / ages_array.length
 end
